Allows list downloading and tracking
External file that has arguments for importing a list of URL's as well as downloading a list in the queue/cache.
After downloading an image it sets the download flag to true - so if the script breaks in middle it will just pick up from where it left off.

Some measures here are redundent and can be consolidated further. A measure is in place to keep the zoom level within JPEG limits.

Small sleep included so we (hopefully) don't hit 429 errors.