Add embed metadata, and revise filename
Revising the filename for sorting on artist, and chronological order. While author name is usually in the URL along with the art name, it's impossible to separate the two by code. Hence reliance on metadata tags to grab author (and date of painting). Would need a backup plan when it can't find author in the data - though it shouls still be named fine. Cases where author name is not in the URL, however, may cause incorrectly truncating the image name.

The embed metadata can be viewed by dropping the final image into exiftool(-k) and so forth.